openspending/openspending#999 Move all services to one folder
