+++ app/controllers/settings.js
@@ -2,7 +2,9 @@
 
 module.exports.main = function(req, res) {
   var config = req.app.get('config');
-  res.json({api: config.get('api'),
-            search: config.get('search')});
+  res.json({
+    api: config.get('api'),
+    search: config.get('search')
+  });
 };
