--- app/front/scripts/components/index.js
@@ -1,8 +1,0 @@
-'use strict';
-
-module.exports.osViewerService = require('./os-viewer-service');
-module.exports.downloader = require('./downloader');
-module.exports.babbage = require('babbage.ui/lib/bindings/angular');
-module.exports.dataPackageApi = require('./data-package-api');
-module.exports.papaparse = require('papaparse');
-
