+++ app/front/scripts/services/index.js
@@ -4,3 +4,7 @@ require('./history');
 require('./login');
 require('./navigation');
 require('./settings');
+
+module.exports.osViewerService = require('./os-viewer-service');
+module.exports.downloader = require('./downloader');
+module.exports.dataPackageApi = require('./data-package-api');
