+++ app/front/scripts/services/navigation/index.js
