+++ app/front/scripts/services/os-viewer-service/index.js
